requirements.txt
streamlit
langchain
langgraph
cassio
langchain_community
tiktoken
langchain-openai
langchainhub
langchain_huggingface
wikipedia
beautifulsoup4
requests
pillow
matplotlib
networkx